moves checkLevelWinCondition to its own file
